chore: deprecate service_name
It is better to subclass the authenticator of interest and set the
login_service property on the subclass.

Support for service_name will be removed in a feature version.
